Finish function for "frequentist-friendly" draws
Finished writing and testing draw_bounds, which makes posterior draws
for the boundaries of the identified set, both with and without
restrictions. For the moment we do not have an analytical formula for
calculating the rho_uz bounds under restrictions on kappa and rho_T*u so
the function just fills in NAs. Once we have this formula, we will need
to write functions to implement it and call them from draw_bounds. The
next step from here is to write a function that summarizes the draws
generated by draw_bounds in terms of medians, HPDIs, etc.