Add code to handle empty identified sets
This commit changes the function draw_bounds so that it returns an NA for any bounds that impose a user-supplied restriction on kappa that leads to an
empty identified set. It is important that we return an NA rather than
an error because we want to be able to calculate the probability that a
particular restriction leads to an empty identified set.

If the user specifies a lower bound for kappa that is *smaller* than the
lower bound for kappa that we calculate from the observables we simply
use the tighter bound that comes from the data. In contrast, when the user specifies an *upper* bound for kappa this always imposes a restriction
on the problem. This is because the data alone give no upper bound for kappa other than the trivial bound (kappa <= 1). In some cases, however, the user-specified upper bound could be *smaller* than the *lower* bound for kappa that we calculate from the data. When this happens, the identified set is empty. This can indeed happen in practice. In the Colonial Origins example from the paper we begin by imposing kappa < 0.6. In this case we find that  27% of the draws for the observables lead to lower bounds for kappa that *exceed* 0.6.