Fix typo in draw_bounds
The function draw_bounds is supposed to return NA for bounds calculated
based on a user-supplied restriction that leads to an empty identified
set. I initially wrote an inequality in the wrong direction so that it
returned NA when the identified set was *not* empty. This commit fixes
the mistake.