Start work in SummarizeAndPlot.R
This file will contain functions for plotting and summarizing results.
This commit moves get_estimates from IdentifiedSet.R to
SummarizeAndPlot.R since it makes more sense for it to be there. It also
adds a function get_HPDI for calculating the posterior median along with
a 90% HPDI using the package coda.